Step A1: port BattleData slot-1 repack from batched branch
All per-turn-mutable fields (winner/flags/activeMonIndex/lastExecTs/turnId) packed into slot 1
so per-turn BattleData mutations coalesce to one SSTORE (main wrote 2: turnId in slot 0).
turnId uint64->uint16, lastExecuteTimestamp uint48->uint40. Engine field access unchanged
(Solidity handles the new slots). Verified: EngineTest 50/50, InlineEngineGasTest 3/3.

+++ src/Engine.sol
@@ -728,7 +728,7 @@ contract Engine is IEngine, MappingAllocator {
             config.p0Move.packedMoveIndex = 0;
             config.p1Move.packedMoveIndex = 0;
         }
-        battle.lastExecuteTimestamp = uint48(block.timestamp);
+        battle.lastExecuteTimestamp = uint40(block.timestamp);
 
         emit EngineExecute(battleKey);
     }
